tests/wpt/web-platform-tests/2dcontext/the-canvas-state/2d.state.saverestore.underflow.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.state.saverestore.underflow</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.state.saverestore.underflow</h1>
<p class="desc">restore() with an empty stack has no effect</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("restore() with an empty stack has no effect");
_addTest(function(canvas, ctx) {

for (var i = 0; i < 16; ++i)
    ctx.restore();
ctx.lineWidth = 0.5;
ctx.restore();
_assertSame(ctx.lineWidth, 0.5, "ctx.lineWidth", "0.5");


});
</script>
